Remove CDN version of jQuery.
Switched to using a locally hosted jQuery exclusively, instead of trying
the CDN first and then falling back to local. The timeout on the CDN
requests kills performance and can result in failure of the app.

See issue #23 for an example of a failed jQuery load causing the
display of raw JSON results behind China's CDN-denying firewall.